# docs/swagger.yaml
definitions:
...
  models.Kelas:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nama:
        type: string
      sekolah_id:
        type: integer
      tahun_ajaran_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  sekolahController.InputKelas:
    properties:
      nama:
        type: string
      sekolah_id:
        type: integer
      tahun_ajaran_id:
        type: integer
    required:
    - nama
    - sekolah_id
    - tahun_ajaran_id
    type: object
...
  sekolahController.UpdateInputKelas:
    properties:
      nama:
        type: string
    required:
    - nama
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /sekolah/kelas:
    post:
      description: Membuat data Kelas baru.
      parameters:
      - description: the body to create a new Kelas
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.InputKelas'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kelas'
      security:
      - BearerToken: []
      summary: Membuat data Kelas barun.
      tags:
      - Sekolah -> Kelas
  /sekolah/kelas/{id}:
    delete:
      description: Menghapus Sebuah data Kelas  berdasarkan id.
      parameters:
      - description: Kelas id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Menghapus data Kelas.
      tags:
      - Sekolah -> Kelas
    get:
      description: Mendapatkan data kelas berdasarkan id.
      parameters:
      - description: kelas id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kelas'
      summary: Mendapatkan data kelas.
      tags:
      - Sekolah -> Kelas
    patch:
      description: memperbaharui data Kelas berdasarkan id.
      parameters:
      - description: Kelas id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Kelas
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.UpdateInputKelas'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kelas'
      security:
      - BearerToken: []
      summary: Memperbarui data Kelas.
      tags:
      - Sekolah -> Kelas
...
  /sekolah/tahun-ajaran/{id}/kelas:
    get:
      description: Mendapatkan semua data kelas Berdasarkan ID Tahun Ajaran.
      parameters:
      - description: Tahun ajaran id  id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Kelas'
            type: array
      summary: mendapatkan data kelas berdasarkan ID Tahun Ajaran.
      tags:
      - Sekolah -> Kelas
...
swagger: "2.0"